fix data type bug and content-type header

+++ include/api_client.h
@@ -7,7 +7,7 @@
 #define MAX_HTTP_OUTPUT_BUFFER 1024
 
 esp_err_t client_event_handler(esp_http_client_event_t *event);
-uint8_t api_get(char **content, char *auth_token, char *device_key, char *endpoint);
-uint8_t api_post(char **content, char *auth_token, char *device_key, char *endpoint, char *post_data);
+uint16_t api_get(char **content, char *auth_token, char *device_key, char *endpoint);
+uint16_t api_post(char **content, char *auth_token, char *device_key, char *endpoint, char *post_data);
 
 #endif
